new snapshot (fix design problem; should now basically work)
The "current position" (ie. the step between last step
and next step) does indeed exist and must be taken
into account; but it cannot be stored in the History class
because it might change (through cursor movement
that does not qualify as separate step). So the "current
step" is always directly requested from TextBuffer
and supplied to History object.